Module rtic_core::multibin

Functions

multibin_cfg_core

If multibin feature is enabled, this returns a tokenstream for the attribute #[cfg(core = "x")] to partition an application to multiple binaries. Otherwise None is returned

multibin_cfg_not_core

If multibin feature is enabled, this returns a tokenstream for the attribute #[cfg(not(core = "x"))] Otherwise None is returned

multibin_shared

If multibin feature is enabled, this returns a tokenstream for the attribute #[multibin_shared] to make sure the annotated variable is present at the same address on all cores. Otherwise None is returned