Crate rtic_core

Core abstractions of the Real-Time Interrupt-driven Concurrency (RTIC) framework

You can write generic libraries using the Mutex trait in this crate. If you want to write application code then you’ll need an implementation of the RTIC framework for a particular architecture. Currently, there are implementations for these architectures and OSes:

ARM Cortex-M

Modules

prelude

Makes locks work on N-tuples, locks the mutexes from left-to-right in the tuple. These are used to reduce rightward drift in code and to help make intentions clearer.

Structs

Exclusive

Newtype over &'a mut T that implements the Mutex trait

Traits

Mutex

Memory safe access to shared resources