Crate rtic_core

TODO: Crate level documentation needed to describe the following:

a distribution

a compilation pass

how compilation passes work

the built-in core compilation pass

guidelines for implementing new distributions, links, and link to template distribution

Modules

multibin

parse_utils

Provides a utility to streamline parsing and manipulating and reconstructing the tokenstream representation of the #[app(arg1=“val1”, …)] attribute

rtic_functions

rtic_traits

Structs

Analysis

App

AppArgs

Arguments provided to the #[app(…)] macro attribute, this includes paths to PACs, number of cores, and peripherals option.

RticMacroBuilder

This should be used to compose an RTIC distribution. In other words, it allows building the RTIC app macro By providing the necessary low-level hardware bindings and binding additional Compilation Passes in the case syntax extensions are desired.

SubAnalysis

SubApp

Traits

CorePassBackend

Interface for providing the low-level hardware bindings specific for a target(s) (A.k.a The Backend) to be used during code generation phase of the *Core Compilation Pass.

RticPass

A trait that allows defining a Compilation Pass.