Trait rtic_core::Mutex

pub trait Mutex {
    type T;
    fn lock<R>(&mut self, f: impl FnOnce(&mut Self::T) -> R) -> R;
}

Memory safe access to shared resources

In RTIC, locks are implemented as critical sections that prevent other tasks from starting. These critical sections are implemented by temporarily increasing the dynamic priority of the current context. Entering and leaving these critical sections is always done in bounded constant time (a few instructions in bare metal contexts).

Associated Types

type T

Data protected by the mutex

Required methods

fn lock<R>(&mut self, f: impl FnOnce(&mut Self::T) -> R) -> R

Creates a critical section and grants temporary access to the protected data